Adding score updating

+++ client/src/state/Player.ts
@@ -16,5 +16,5 @@ export class Player extends Schema {
     @type("int8") public index: number;
     @type("boolean") public isHost: boolean;
     @type("boolean") public moving: boolean;
-    @type("int8") public score: number;
+    @type("int32") public score: number;
 }
